Abstract out the voting models.
Before, we had only a single voting model that was tightly coupled to the User. This creates the notion of a voting model which can be added to a User, allowing easier changes to different voting models, as well as at some point allow different populations with different voting models.

Ultimately, this is a no-op for population_builder.py.